# backend/internal/docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: CRM Lite API
  version: 2.0.0
  description: 'Portfolio CRM REST API (Go/Gin + PostgreSQL + Redis).


    All JSON endpoints return the shared envelope:


    ```json

    { "success": true|false, "message": "...", "data": {}, "errors": [] }

    ```


    Authenticate with `Authorization: Bearer <access_token>` from `POST /auth/login`.

    Organization scope is resolved server-side from the token (tenant middleware).

    '
servers:
- url: __API_SERVER_URL__
  description: Current host (injected at request time)
...
paths:
...
  /attachments/{attachmentId}:
    delete:
      summary: Delete attachment
      tags:
      - Attachments
      operationId: deleteAttachment
      security:
      - bearerAuth: []
      parameters:
      - name: attachmentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /modules/{id}/fields/{fieldId}:
    get:
      summary: Get field
      tags:
      - Fields
      operationId: getField
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      - name: fieldId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/APIResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/FieldResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update field
      tags:
      - Fields
      operationId: updateField
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      - name: fieldId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFieldRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/APIResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/FieldResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
    delete:
      summary: Delete field
      tags:
      - Fields
      operationId: deleteField
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      - name: fieldId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from POST /auth/login
  responses:
    BadRequest:
      description: Malformed request or business validation failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            simple:
              value:
                success: false
                message: Invalid request
                errors: null
            fieldErrors:
              value:
                success: false
                message: Validation failed
                errors:
                - field: email
                  message: Must be a valid email
            appError:
              value:
                success: false
                message: Invalid request
                errors:
                - code: BAD_REQUEST
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Unauthorized
            errors: null
    Forbidden:
      description: Authenticated but lacking permission / module ACL
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Forbidden
            errors: null
    NotFound:
      description: Resource not found in the caller's organization
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Not found
            errors:
            - code: NOT_FOUND
...
  schemas:
    APIResponse:
      type: object
      required:
      - success
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: OK
        data:
          description: Endpoint-specific payload
          nullable: true
        errors:
          nullable: true
          description: Field errors or app error items when success is false
          oneOf:
          - type: array
            items:
              $ref: '#/components/schemas/FieldValidationError'
          - type: array
            items:
              $ref: '#/components/schemas/AppErrorItem'
    ErrorResponse:
      type: object
      required:
      - success
      - message
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation failed
        data:
          nullable: true
        errors:
          nullable: true
          oneOf:
          - type: array
            items:
              $ref: '#/components/schemas/FieldValidationError'
          - type: array
            items:
              $ref: '#/components/schemas/AppErrorItem'
    ErrorCode:
      type: string
      enum:
      - BAD_REQUEST
      - UNAUTHORIZED
      - FORBIDDEN
      - NOT_FOUND
      - CONFLICT
      - VALIDATION_ERROR
      - INTERNAL_SERVER_ERROR
      example: BAD_REQUEST
    FieldValidationError:
      type: object
      required:
      - field
      - message
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: Must be a valid email
    AppErrorItem:
      type: object
      required:
      - code
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
...
    FieldOption:
      type: object
      required:
      - label
      - value
      properties:
        label:
          type: string
          example: Hot
        value:
          type: string
          example: hot
    FieldResponse:
      type: object
      properties:
        id: *id002
        module_id:
          type: string
          format: uuid
        api_name:
          type: string
          example: amount
        label:
          type: string
          example: Amount
        field_type:
          type: string
          enum:
          - text
          - textarea
          - email
          - phone
          - number
          - currency
          - date
          - datetime
          - boolean
          - dropdown
          - multiselect
          - radio
          - checkbox
          - url
          - file
          - image
          - user
          - lookup
          - formula
          - json
          - richtext
        is_required:
          type: boolean
        is_unique:
          type: boolean
        is_read_only:
          type: boolean
        is_visible:
          type: boolean
        is_searchable:
          type: boolean
        is_filterable:
          type: boolean
        is_nullable:
          type: boolean
        is_indexed:
          type: boolean
        is_system:
          type: boolean
        default_value:
          type: string
          nullable: true
        placeholder:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        help_text:
          type: string
          nullable: true
        regex:
          type: string
          nullable: true
        validation_message:
          type: string
          nullable: true
        lookup_module_id:
          type: string
          format: uuid
          nullable: true
        min_length:
          type: integer
          nullable: true
        max_length:
          type: integer
          nullable: true
        sort_order:
          type: integer
        options:
          type: array
          items:
            $ref: '#/components/schemas/FieldOption'
        storage:
          type: object
          properties:
            kind:
              type: string
              enum:
              - column
              - jsonb
            path:
              type: string
        created_at: *id003
        updated_at: *id003
...
    UpdateFieldRequest:
      type: object
      properties:
        label:
          type: string
        is_required:
          type: boolean
        is_unique:
          type: boolean
        is_read_only:
          type: boolean
        is_visible:
          type: boolean
        is_searchable:
          type: boolean
        is_filterable:
          type: boolean
        default_value:
          type: string
          nullable: true
        placeholder:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        help_text:
          type: string
          nullable: true
        regex:
          type: string
          nullable: true
        validation_message:
          type: string
          nullable: true
        min_length:
          type: integer
          nullable: true
        max_length:
          type: integer
          nullable: true
        sort_order:
          type: integer
        options:
          type: array
          items:
            $ref: '#/components/schemas/FieldOption'
...